Don't search for request in parent scope in groups
Since our current semantics don't properly capture requests sent in a
parent scope, it is only logical to disable searching in parent scope
alltogether, as this forces the user to adopt the exact style of
simplified PUML code that we desire. NOTE that this change does not
limit the permitted Expressiveness, instead only forcing the user to
make the desired semantics explicit by potentially duplicating some
code.