Readd search for request in outer scopes
The only case where this logic "breaks" is with diagrams where the PUML
semantics are highly questionable at best. In such a case we will simply
blame the user that his PUML is wrong and take the extra convenience of
being able to define requests outside alt blocks and corresponding
responses within.